Reorganized - all provisioner stuff into GradleProvisioner, all the rest into RegisterDependenciesTask.
